@@ -1,7 +1,10 @@
 
 package com.polar.nextcloudservices;
 
-import android.content.SharedPreferences;
+import static androidx.core.app.NotificationCompat.PRIORITY_MIN;
+import static com.polar.nextcloudservices.Preferences.PreferencesUtils.NONE_RESULT;
+import static com.polar.nextcloudservices.ui.settings.SettingsFragment.SSO_ENABLED_PREFERENCE;
+
 import android.os.Build;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
@@ -18,7 +21,7 @@ import android.net.NetworkInfo;
 
 
 import androidx.core.app.NotificationCompat;
-import androidx.preference.PreferenceManager;
+import androidx.core.app.NotificationManagerCompat;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -37,13 +40,18 @@ import com.nextcloud.android.sso.model.SingleSignOnAccount;
 import com.polar.nextcloudservices.NotificationProcessors.BasicNotificationProcessor;
 import com.polar.nextcloudservices.NotificationProcessors.NextcloudTalkProcessor;
 import com.polar.nextcloudservices.NotificationProcessors.OpenBrowserProcessor;
+import com.polar.nextcloudservices.Preferences.PreferencesUtils;
 
 class PollTask extends AsyncTask<NotificationService, Void, JSONObject> {
     private final String TAG = "NotifcationService.PollTask";
+    private Context mContext;
 
+    public PollTask (Context context){
+        mContext = context;
+    }
     @Override
     protected JSONObject doInBackground(NotificationService... services) {
-        return services[0].API.getNotifications(services[0]);
+        return services[0].API.getNotifications(services[0], mContext);
     }
 }
 
@@ -51,10 +59,16 @@ public class NotificationService extends Service {
     // constant
     public long pollingInterval = 3 * 1000; // 3 seconds
     public static final String TAG = "NotificationService";
+
+    private static final String BACKGROUND_NOTIFICATION_CHANNEL = "__internal_backgorund_polling";
+    private static final String SERVERCONFIG_NOTIFICATION_CHANNEL = "SERVERCONFIG_NOTIFICATION_CHANNEL";
+
+    private static final int SERVERCONFIG_NOTIFICATION_CHANNEL_ID = 5456161;
+
+
     public String server = "";
     public String username = "";
     public String password = "";
-    public String status = "Disconnected";
     public boolean useHttp = false;
     public boolean allowRoaming = false;
     public boolean allowMetered = false;
@@ -62,6 +76,16 @@ public class NotificationService extends Service {
     private PollTimerTask task;
     public NextcloudAbstractAPI API;
     private NotificationBuilder mNotificationBuilder;
+
+    private String mStatusReason = "Disconnected";
+    private STATE mStatus = STATE.DISCONNECTED;
+
+    public enum STATE {
+        CONNECTED,
+        DISCONNECTED,
+        DISABLED
+    }
+
 
     private void registerNotificationProcessors(){
         if(mNotificationBuilder==null){
@@ -81,7 +105,7 @@ public class NotificationService extends Service {
 
         @Override
         public void onError(Exception ex) {
-            status = "Disconnected: " + ex.getLocalizedMessage();
+            setStatus(STATE.DISCONNECTED, ex.getLocalizedMessage());
             ex.printStackTrace();
         }
     };
@@ -92,8 +116,17 @@ public class NotificationService extends Service {
     // timer handling
     private Timer mTimer = null;
 
-    public String getStatus() {
-        return this.status;
+    public STATE getStatus() {
+        return this.mStatus;
+    }
+
+    public String getStatusReason() {
+        return this.mStatusReason;
+    }
+
+    public void setStatus(STATE stat, String reason){
+        mStatus = stat;
+        mStatusReason =  reason;
     }
 
     public boolean checkInternetConnection(Context context) {
@@ -153,10 +186,10 @@ public class NotificationService extends Service {
                     mNotificationManager.cancel(remove_id);
                     active_notifications.remove(remove_id);
                 }
-                this.status = "Connected";
+                setStatus(STATE.CONNECTED, "Unknown");
 
             } catch (Exception e) {
-                this.status = "Disconnected: " + e.getLocalizedMessage();
+                setStatus(STATE.DISCONNECTED, e.getLocalizedMessage());
                 e.printStackTrace();
             }
         }
@@ -177,6 +210,13 @@ public class NotificationService extends Service {
 
     @Override
     public void onCreate() {
+
+        if(!verifyValidPreferences(server)){
+            Log.e(TAG, "The Serverconfiguration is invalid!");
+            createInvalidConfigNotification();
+            return;
+        }
+
         // cancel if already existed
         if (mTimer != null) {
             mTimer.cancel();
@@ -188,25 +228,26 @@ public class NotificationService extends Service {
         task = new PollTimerTask();
         mTimer.scheduleAtFixedRate(task, 0, pollingInterval);
         //Create background service notifcation
-        String channelId = "__internal_backgorund_polling";
         NotificationManager mNotificationManager =
                 (NotificationManager) getSystemService(NOTIFICATION_SERVICE);
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            NotificationChannel channel = new NotificationChannel(channelId, "Background polling", NotificationManager.IMPORTANCE_LOW);
+            NotificationChannel channel = new NotificationChannel(
+                    BACKGROUND_NOTIFICATION_CHANNEL,
+                    getString(R.string.notification_persistent_title),
+                    NotificationManager.IMPORTANCE_LOW);
             mNotificationManager.createNotificationChannel(channel);
         }
         //Build notification
         NotificationCompat.Builder mBuilder =
-                new NotificationCompat.Builder(this, channelId)
+                new NotificationCompat.Builder(this, BACKGROUND_NOTIFICATION_CHANNEL)
                         .setSmallIcon(R.drawable.ic_logo)
                         .setContentTitle(getString(R.string.app_name))
-                        .setPriority(-2)
+                        .setPriority(PRIORITY_MIN)
                         .setOnlyAlertOnce(true)
-                        .setContentText("Background connection notification");
+                        .setContentText(getString(R.string.notification_persistent_content));
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            mBuilder.setChannelId(channelId);
+            mBuilder.setChannelId(BACKGROUND_NOTIFICATION_CHANNEL);
         }
-
 
         Notification mNotification = mBuilder.build();
         //Here we want to get Nextcloud account if it does exist
@@ -219,13 +260,24 @@ public class NotificationService extends Service {
         registerNotificationProcessors();
     }
 
+    private boolean verifyValidPreferences(String URL) {
+        if(PreferencesUtils.getBoolPreference(this, SSO_ENABLED_PREFERENCE, false)){
+            return true;
+        }
+        if(URL.equals(NONE_RESULT)){
+            return false;
+        }
+        return true;
+    }
+
     private void updateAccounts() {
-        if (getBoolPreference("sso_enabled", false)) {
-            final String name = getPreference("sso_name");
-            final String server = getPreference("sso_server");
-            final String type = getPreference("sso_type");
-            final String token = getPreference("sso_token");
-            final String userId = getPreference("sso_userid");
+        if (PreferencesUtils.getBoolPreference(this, SSO_ENABLED_PREFERENCE, false)) {
+            //todo: move keys to constants
+            final String name = PreferencesUtils.getPreference(this, "sso_name");
+            final String server = PreferencesUtils.getPreference(this, "sso_server");
+            final String type = PreferencesUtils.getPreference(this, "sso_type");
+            final String token = PreferencesUtils.getPreference(this, "sso_token");
+            final String userId = PreferencesUtils.getPreference(this, "sso_userid");
             final SingleSignOnAccount ssoAccount = new SingleSignOnAccount(name, userId, token, server, type);
             NextcloudAPI mNextcloudAPI = new NextcloudAPI(this, ssoAccount, new GsonBuilder().create(), apiCallback);
             API = new NextcloudSSOAPI(mNextcloudAPI);
@@ -237,23 +289,8 @@ public class NotificationService extends Service {
         }
     }
 
-    private String getPreference(String key) {
-        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
-        return sharedPreferences.getString(key, "<none>");
-    }
-
-    private Integer getIntPreference(String key) {
-        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
-        return sharedPreferences.getInt(key, Integer.MIN_VALUE);
-    }
-
-    private boolean getBoolPreference(String key, boolean fallback) {
-        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
-        return sharedPreferences.getBoolean(key, fallback);
-    }
-
     public void onPreferencesChange() {
-        int _pollingInterval = getIntPreference("polling_interval") * 1000;
+        int _pollingInterval = PreferencesUtils.getIntPreference(this, "polling_interval") * 1000;
         if (_pollingInterval <= 0) {
             Log.w(TAG, "Invalid polling interval! Setting to 3 seconds.");
             _pollingInterval = 3 * 1000;
@@ -264,7 +301,6 @@ public class NotificationService extends Service {
             pollingInterval = _pollingInterval;
             updateTimer();
         }
-
     }
 
     @Override
@@ -274,21 +310,31 @@ public class NotificationService extends Service {
         mTimer.purge();
     }
 
+    private void createInvalidConfigNotification(){
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            int importance = NotificationManager.IMPORTANCE_DEFAULT;
+            NotificationChannel channel = new NotificationChannel(SERVERCONFIG_NOTIFICATION_CHANNEL, getString(R.string.serverconfig_notification_channel_name), importance);
+            channel.setDescription(getString(R.string.notification_channel_invalid_config_description));
+            // Register the channel with the system; you can't change the importance
+            // or other notification behaviors after this
+            NotificationManager notificationManager = getSystemService(NotificationManager.class);
+            notificationManager.createNotificationChannel(channel);
+        }
+
+        NotificationCompat.Builder builder = new NotificationCompat.Builder(this, SERVERCONFIG_NOTIFICATION_CHANNEL)
+                .setSmallIcon(R.drawable.ic_logo)
+                .setContentTitle(getString(R.string.serverconfig_notification_wrong_config))
+                .setContentText(getString(R.string.serverconfig_notification_wrong_config_long))
+                .setPriority(NotificationCompat.PRIORITY_DEFAULT);
+
+        NotificationManagerCompat notificationManager = NotificationManagerCompat.from(this);
+        notificationManager.notify(SERVERCONFIG_NOTIFICATION_CHANNEL_ID, builder.build());
+    }
 
     public class Binder extends android.os.Binder {
         // Returns current status string of a service
-        public String getServiceStatus() {
+        public STATE getServiceStatus() {
             return getStatus();
-        }
-
-        // Runs re-check of preferences, can be called from activities
-        public void onPreferencesChanged() {
-            onPreferencesChange();
-        }
-
-        // Update API class when accounts state change
-        public void onAccountChanged() {
-            updateAccounts();
         }
     }
 
@@ -297,37 +343,23 @@ public class NotificationService extends Service {
         @Override
         public void run() {
             // run on another thread
-            mHandler.post(new Runnable() {
+            mHandler.post(() -> {
+                username = PreferencesUtils.getPreference(getApplicationContext(), "login");
+                password = PreferencesUtils.getPreference(getApplicationContext(), "password");
+                server = PreferencesUtils.getPreference(getApplicationContext(), "server");
+                useHttp = PreferencesUtils.getBoolPreference(getApplicationContext(), "insecure_connection", false);
+                allowRoaming = PreferencesUtils.getBoolPreference(getApplicationContext(), "allow_roaming", false);
+                allowMetered = PreferencesUtils.getBoolPreference(getApplicationContext(), "allow_metered", false);
 
-                @Override
-                public void run() {
-                    username = getPreference("login");
-                    password = getPreference("password");
-                    server = getPreference("server");
-                    useHttp = getBoolPreference("insecure_connection", false);
-                    allowRoaming = getBoolPreference("allow_roaming", false);
-                    allowMetered = getBoolPreference("allow_metered", false);
+                //FIXME: Should call below method only when prefernces updated
+                onPreferencesChange();
 
-
-                    //FIXME: Should call below method only when prefernces updated
-                    onPreferencesChange();
-
-                    if (checkInternetConnection(getApplicationContext())) {
-                        new PollTask().execute(NotificationService.this);
-                    } else {
-                        status = "Disconnected: no network available";
-                    }
+                if (checkInternetConnection(getApplicationContext())) {
+                    new PollTask(getApplicationContext()).execute(NotificationService.this);
+                } else {
+                    setStatus(STATE.DISCONNECTED, getString(R.string.state_noconnection));
                 }
-
             });
         }
-
-        private String getDateTime() {
-            // get date time in custom format
-            SimpleDateFormat sdf = new SimpleDateFormat("[yyyy/MM/dd - HH:mm:ss]");
-            return sdf.format(new Date());
-        }
-
-
     }
 }
