@@ -1,23 +1,34 @@
-package com.polar.nextcloudservices;
+package com.polar.nextcloudservices.ui.credits;
 
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
 import android.util.Log;
+import android.view.LayoutInflater;
 import android.view.View;
+import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ListView;
+import android.widget.TextView;
 
-import androidx.appcompat.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProvider;
 
 import com.polar.nextcloudservices.Adapters.CreditsAdapter;
+import com.polar.nextcloudservices.ContributorDetails;
+import com.polar.nextcloudservices.R;
+import com.polar.nextcloudservices.databinding.FragmentCreditsBinding;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class CreditsActivity extends AppCompatActivity {
+public class CreditsFragment extends Fragment {
 
-    private final String TAG = "CreditsActivity";
+
+    private final String TAG = "CreditsFragment";
     List<ContributorDetails> details = new ArrayList<>();
     String[] licenses;
     String[] urls;
@@ -25,11 +36,20 @@ public class CreditsActivity extends AppCompatActivity {
     String[] owner_github_Name;
     String[] owner_github_image;
 
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    private CreditsViewModel creditsViewModel;
+    private FragmentCreditsBinding binding;
+
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        creditsViewModel =
+                new ViewModelProvider(this).get(CreditsViewModel.class);
+
+        binding = FragmentCreditsBinding.inflate(inflater, container, false);
+        View root = binding.getRoot();
+
         super.onCreate(savedInstanceState);
-        setTitle("Credits");
-        setContentView(R.layout.activity_credits);
+        //setTitle("Credits");
+        //setContentView(R.layout.activity_credits);
         licenses = getResources().getStringArray(R.array.oss_libs);
         urls = getResources().getStringArray(R.array.oss_libs_links);
         owner_Name = getResources().getStringArray(R.array.oss_libs_owner_name);
@@ -37,12 +57,10 @@ public class CreditsActivity extends AppCompatActivity {
         owner_github_image = getResources().getStringArray(R.array.oss_libs_owner_Img);
 
         details_add();
-        ListView mListView = (ListView) findViewById(R.id.oss_licenses_list);
-        if (mListView == null) {
-            Log.wtf(TAG, "ListView is null!");
-            throw new RuntimeException("ListView should not be null!");
-        }
-        CreditsAdapter aAdapter = new CreditsAdapter(this, R.layout.credits_contributer, details);
+
+
+        ListView mListView = binding.ossLicensesList;
+        CreditsAdapter aAdapter = new CreditsAdapter(this.getContext(), R.layout.credits_contributer, details);
         mListView.setAdapter(aAdapter);
         mListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
 
@@ -51,6 +69,14 @@ public class CreditsActivity extends AppCompatActivity {
                 startActivity(browserIntent);
             }
         });
+
+        return root;
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        binding = null;
     }
 
     //    A function to add the contributor details in details(Arraylist) which is to be shown in credits activity
@@ -59,5 +85,4 @@ public class CreditsActivity extends AppCompatActivity {
             details.add(new ContributorDetails(owner_Name[i], licenses[i], owner_github_image[i], owner_github_Name[i]));
         }
     }
-
 }
